API generate with Solana program call (#37)

+++ src/models/conversion_operation_config_params.rs
@@ -27,7 +27,7 @@ pub struct ConversionOperationConfigParams {
         rename = "slippageBasisPoints",
         skip_serializing_if = "Option::is_none"
     )]
-    pub slippage_basis_points: Option<u32>,
+    pub slippage_basis_points: Option<u16>,
 }
 
 impl ConversionOperationConfigParams {
